Plugin architecture, HTTP API, SITL mission execution
This commit refactors the app workflow to get the basics of a plugin-type workflow happening, in the app/routines directory.  Build system is updated to compose assets from all subdirs there (templates, etc) and the app/routines/plugin dir is an example for "routine" or "mission" template.

The idea is that code in the /app/routines directory is invoked dynamically based on the user's choice of platform/payload/mission type, and that plugin can use some "global" widgets but implement its own GUI as desired.

The GCS server now accepts certain HTTP endpoints to trigger certain actions, like starting to listen for a connection, etc.  It's a work in progress because it's not entirely clear what the API should look like.