Closes #100, handle user-defined params from planning
This commit sketches code between the client GUI/server side to manage UAV parameters interactively.  The branch started to handle max speed params, but we changed course to just touch lightly on both that param as well as handling the maxAltitude simply via GUI/server controls.

This branch doesn't add a lot to the core code, but contains some sketches that are useful for further work when we make these things more complete.