Fixes #167, adds SITL as platform
Also adds guard code that prevents the GCS from sending updated params to the platform if none are defined in the assets/js/libs/platforms.js file.  Reason being we always need to provision the SITL with params, but not always on other real platforms where we may want to manage the param set differently.