Progress towards coherent shared state, #202, #200, #198, #197, #187, #186, #185, #179.
Server-side "Users" module starts to aggregrate the changes needed to manage connected users, including who's currently promoted as operator.  Total global state management structure is evolving a bit, because it's necessary to listen in a few places for whether a mission has started, etc, and/or how to render controls on the Fly screen, or how to promote the GUI so someone who becomes operator can now operate the craft.  Somewhat bigger changes than I'd hoped for, but good stuff.