Issues #205, #203, #186, #187, #118
Improves handling with multiple clients connecting/disconnecting (work remains to ensure mutex on operator, as well as timing of when routines/operators are assigned).  Navigation between Mission and Engineering views is working now.  Issue with blank tiles showing is repaired by doing a more careful check of when GPS fix is established.  Clients are now notified when they are demoted.