Addresses #240, #302, use MAV_CMD_NAV_LOITER_UNLIM
This isn't a fix or a close for either issue, but it's related to both.  Attempt here was to try and improve the "lost command" thing when the user clicks/releases too quickly.  I think the likely fix needs to be clever handling/queueing of messages being sent.  See #302 for details.